internal module rearrangement
