+++ calval/providers/landsat/__init__.py
@@ -9,12 +9,12 @@ import dateutil.parser
 import numpy as np
 from calval.geometry import IncidenceAngle
 from calval.normalized_scene import band_names
-from calval.scene_info import SceneInfo, extract_archive, scaling
-from calval.scene_data import SceneData
-from calval.landsat_mtl import read_mtl, ephemeris_df
 from calval.sun_locator import SunLocator
 from calval.satellites.srf import Landsat8Blue, Landsat8Green, Landsat8Red, Landsat8Nir
 from calval.analysis import srf_exatmospheric_irradiance
+from calval.providers import SceneData, SceneInfo
+from calval.providers.scene_info import extract_archive, scaling
+from .landsat_mtl import read_mtl, ephemeris_df
 
 _site_prs = {
     'baotou': ['128032', '127032'],
