+++ calval/providers/scene_info.py
