+++ calval/utils/batch_plot.py
@@ -2,7 +2,7 @@
 This module is a hack in order to avoid errors in case of batch plotting (e.g. in tests),
 where there is no interactive environment properly configured (e.g. in containers).
 
-Usage: `from calval.batch_plot import plt`
+Usage: `from calval.utils.batch_plot import plt`
 Make sure this line comes before importing any third-party package which depends on
 `matplotlib.pyplot`
