+++ make_sm.py
@@ -8,13 +8,12 @@ from tqdm import tqdm
 from azure.storage.blob import BlockBlobService
 
 from calval.normalized_scene import band_names
-from calval.scene_info import SceneInfo
+from calval.providers import SceneInfo, SceneData
 # Import provider module to enable the factory mechanism
-import calval.sentinel_scenes  # noqa: F401
-import calval.landsat_scenes  # noqa: F401
+import calval.providers.sentinel  # noqa: F401
+import calval.providers.landsat  # noqa: F401
 from calval.scene_utils import make_sat_measurements
 from calval.sat_measurements import SatMeasurements
-from calval.scene_data import SceneData
 
 
 logger = logging.getLogger()
