+++ tests/testing_utils.py
@@ -1,10 +1,9 @@
 import os
 import calval.config
-from calval.scene_info import SceneInfo
-from calval.scene_data import SceneData
+from calval.providers import SceneInfo, SceneData
 # Import provider module to enable the factory mechanism
-import calval.sentinel_scenes  # noqa: F401
-import calval.landsat_scenes  # noqa: F401
+import calval.providers.sentinel  # noqa: F401
+import calval.providers.landsat  # noqa: F401
 
 testdir = os.path.abspath(os.path.dirname(__file__))
 calval.config.shapes_dir = os.path.join(testdir, 'data', 'sites')
